requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
attrs==17.4.0             # via pytest
coverage==4.5             # via pytest-cov
django-config-models==0.1.9
djangorestframework==3.6.4  # via django-config-models
pluggy==0.6.0             # via pytest
py==1.5.2                 # via pytest
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.4.0             # via pytest-cov, pytest-django
pytz==2017.3              # via django
six==1.11.0               # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pytest
